Update podcast.py to include thumbnail generation once every 10 minutes
This should allow for the preset 'snapshots' to be selected, so that a snapshot is taken once every X frames (every 10 minutes for the moment) and stored in the appropriate directory.

And let me take a moment to say, I am thoroughly enjoying github for this as I can do a quick edit like this in the browser without even having to clone the repo.  Works well for small things.